requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
fonttools==4.39.4
frozendict==2.3.8
gunicorn==20.0.4
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multitasking==0.0.11
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scikit-metrics==0.1.0
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
soupsieve==2.4.1
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.2.2
yfinance==0.2.24